* Icy let you select appropriate importer when multiple choice are available. * Loader: many changes to provide better support on FileImporter and SequenceFileImporter classes. * FileImporter.acceptFile(..) method now use path parameter. * Layers panel:  - opacity setting has moved at bottom.  - support custom setting panel per layer (Overlay.getOptionsPanel()) * SequenceUtil:  - C/Z/T merge operation now preserve metadata.  - added getSubSequence(Sequence, ROI, double) method. * User email information stored in preferences. * some refactoring:  - ImageLoaderDialog --> LoaderDialog  - ImageSaverDialog --> SaverDialog  - LoadDialog --> OpenDialog * others minors improvements, bug fixes and cleanup.

+++ icy/type/collection/array/Array1DUtil.java
@@ -295,15 +295,15 @@ public class Array1DUtil
             case BYTE:
                 return Arrays.copyOf((byte[]) array, ((byte[]) array).length);
             case SHORT:
-                return Arrays.copyOf((byte[]) array, ((byte[]) array).length);
+                return Arrays.copyOf((short[]) array, ((short[]) array).length);
             case INT:
-                return Arrays.copyOf((byte[]) array, ((byte[]) array).length);
+                return Arrays.copyOf((int[]) array, ((int[]) array).length);
             case LONG:
-                return Arrays.copyOf((byte[]) array, ((byte[]) array).length);
+                return Arrays.copyOf((long[]) array, ((long[]) array).length);
             case FLOAT:
-                return Arrays.copyOf((byte[]) array, ((byte[]) array).length);
+                return Arrays.copyOf((float[]) array, ((float[]) array).length);
             case DOUBLE:
-                return Arrays.copyOf((byte[]) array, ((byte[]) array).length);
+                return Arrays.copyOf((double[]) array, ((double[]) array).length);
             default:
                 return null;
         }
